split out our wrapping/unwrapping code into its own module
